+++ package.json
@@ -35,6 +35,7 @@
     "@macrostrat/revision-info-webpack": "^1.0.0",
     "@macrostrat/storybook": "workspace:*",
     "@mdx-js/rollup": "^2.3.0",
+    "@types/underscore": "^1",
     "@typescript-eslint/eslint-plugin": "^6.3.0",
     "@typescript-eslint/parser": "^6.3.0",
     "@yarnpkg/sdks": "^3.1.0",
@@ -133,6 +134,7 @@
     "topojson-client": "^3.0.0",
     "transition-hook": "^1.5.2",
     "tsx": "^4.11.2",
+    "underscore": "^1.13.6",
     "use-debounce": "^9.0.4",
     "use-react-router-breadcrumbs": "^3.2.1",
     "use-resize-observer": "^9.1.0",
