+++ src/pages/correlation/+Page.ts
@@ -23,8 +23,7 @@ import { centroid } from "@turf/centroid";
 
 import { buildCrossSectionLayers } from "~/_utils/map-layers";
 import { fetchAllColumns } from "~/pages/map/map-interface/app-state/handlers/fetch";
-import { getHashString, setHashString } from "@macrostrat/ui-components";
-import { fmt2 } from "~/pages/map/map-interface/utils";
+import { getFocusedLineFromHashParams, HashStringManager } from "./hash-string";
 
 interface CorrelationState {
   focusedLine: LineString | null;
@@ -64,76 +63,22 @@ const useCorrelationDiagramStore = create<CorrelationState>((set) => ({
   },
 }));
 
-function getFocusedLineFromHashParams(): LineString | null {
-  if (typeof window === "undefined") {
-    return null;
-  }
-  let hash = getHashString(window.location.hash);
-
-  if (hash?.line == null) {
-    return null;
-  }
-
-  try {
-    let coords = hash.line;
-
-    console.log("coords", coords);
-
-    if (coords.length < 2) {
-      return null;
-    }
-    if (coords.length % 2 != 0) {
-      console.error("Invalid number of coordinates in hash string");
-      return null;
-    }
-
-    coords = chunk(coords, 2);
-
-    return {
-      type: "LineString",
-      coordinates: coords.map((coord) => coord.map(Number)),
-    };
-  } catch (e) {
-    console.error("Error parsing hash string", e);
-    return null;
-  }
-}
-
-function chunk(arr, size) {
-  return Array.from({ length: Math.ceil(arr.length / size) }, (v, i) =>
-    arr.slice(i * size, i * size + size)
-  );
-}
-
 export function Page() {
   const startup = useCorrelationDiagramStore((state) => state.startup);
   useEffect(() => {
     startup();
   }, []);
 
+  const focusedLine = useCorrelationDiagramStore((state) => state.focusedLine);
+
   return h(FullscreenPage, [
-    h(HashStringManager),
+    h(HashStringManager, { focusedLine }),
     h("header", [h(PageBreadcrumbs)]),
     h("div.flex.row", [
       h("div.correlation-diagram"),
       h("div.assistant", [h(InsetMap)]),
     ]),
   ]);
-}
-
-function HashStringManager() {
-  const focusedLine = useCorrelationDiagramStore((state) => state.focusedLine);
-  useEffect(() => {
-    if (focusedLine == null || focusedLine.coordinates.length < 2) {
-      return;
-    }
-    let hash = {
-      line: focusedLine.coordinates.flatMap((coord) => coord.map(fmt2)),
-    };
-    setHashString(hash);
-  }, [focusedLine]);
-
-  return null;
 }
 
 function InsetMap() {
